requirements.txt
pythorhead==0.25.0
PyGithub==2.6.0
requests
smmrpy
aiohttp
pytube
newspaper3k
lxml_html_clean

